+++ packages/redux/src/reducers/reducers.ts
@@ -18,7 +18,6 @@ import {
   itemsReducer,
   navigationReducer,
   breadcrumbsReducer,
-  descriptorsReducer,
   childrenReducer,
   treeReducer
 } from './content';
@@ -26,7 +25,6 @@ import { searchReducer } from './search'
 
 export const allReducers = {
   items: itemsReducer,
-  descriptors: descriptorsReducer,
   children: childrenReducer,
   trees: treeReducer,
   navigation: navigationReducer,